Hipotetical
Before I realized that my definition of a mismatched line was
derailing from the original intent I came up with a clever idea to
solve the issue of the ghost line.

The issue with the Ghost Line is that the strings.Split function
is counting the last line as an empty line ... but that works. I
forgot to run the tests after I modified the tests, so I don't
know if my code passes. From the mental runs I am doing, it should
pass.

Anyway, the problem that I described yester is fuzzy now, so I
don't really recall the issue. Right now as I am reviewing the
problem I see no reason why having an empty string at the end be
an issue, since it is correct. The last line is empty. Again, I
think it has to do with the fact that I switched from thinking
that the newline char signals the end of one line but instead
signals the start of a new one. I mean, it is doing both. It
depends on what angle you look at it. Anyway, the issue was that
it was extracting an empty string from between LFNULL.

To correct this, all you need to do is detect this pattern and
correct the output of Split. That is, after running Split and
getting the slices, you would check if the last char is \n. If it
was, all you would need to do would be to remove the last element
in the resulting slices. And that is it. All the other propositions
I made where overkill.

I like how explicit and simple this solution is (to a problem that
was not a problem). But then I realized that I could do something
more sneaky and efficient. Ignoring the last element of the slice
is equivalent to removing it. At the expense of using more memory,
I could avoid the manipulations of slices by adjusting the max
indices if the condition of interest was met. So that is what I
decided to do, just as an excercise. I had a gut feeling that I
could learn something useful from this pattern. In particular, I
wanted to challenge myself to document/make explicit this approach.
Turns out that my gut feeling was correct: I am learning a lot
about the nuances of working with strings and chars in Go and its
type system.

This solution is a good example of an optimization with marginal
results. Slices are designed so that manipulations like the one I
just described be efficient. Technically speaking, the slice is
going to do the same thing that I am doing it, but internally.
Even though my way would be more 'efficient' the difference is
just a couple of dereferences(of something that is probably in
cache). So yeah, minimal performance improvement, at the expense
of understandability. What I wanted to challenge myself to do is
to attempt to make my intent clear through the use of naming
conventions and comments. Again, just an exercise. The conclusion:
it is not worth it.

But the stuff that I am learning about strings in Go is. Why is
accessing a character in a string a byte? and what is an untyped
string? That last one really threw me off. It implies the
existence of type strings. What is a typed string? This stuff is
weird.