Str to char
I just realized that I was attempting to compare a string to a
char. Clearly that was not going to work.

Again, why is accessing the last char of a string a byte? I recall
from Matt Holiday's lectures on Go that working with strings as if
they were arrays do not result in what you would expect. Not all
characters are stored as bytes. As I understand, due to the use of
UTF-8 some characters might take more than one byte to store. Is
accessing characters using [] and an index safe?

I tried looking for that query in Google and I could not find any
information.

using q="golang string how to access character" I found
https://tech.forums.softwareag.com/t/how-to-index-characters-in-a-golang-string/283123/3
where the answer of Elio Baresi sheds some light on the subject.
"If you need to modify a string, you should convert it to a []rune
(a slice of Unicode code points), make the modifications, and then
convert it back to a string." It is good to know that.

I still had my doubts about what Matt Holiday has mentioned, so I
looked for q="golang how to use length of string" and found
https://www.educative.io/answers/how-to-find-the-length-of-a-string-in-golang
which sheds some light on the subject.

"The len() method returns the number of bytes in a string."

And RuneCountInString returns the number of UTF-8 chars.

"For normal characters, we can use any of the methods. However, if
there are any special characters like the Black Chess Queen ♛, the
len() method returns a length value output which is not accepted."

Does go make a distinction between the two types of strings?
Perhaps that is the reason why there are 'typed' strings in Go.
Like UTFstring and ASCIIstring, and string is the generic(and
unsafe) version. It makes sense because I now find it dreadful
working with strings. How do you check if a string has chars that
use more than one byte?

So the question is, what does [] accesses? a rune or a byte? Well
based on the type signature, it is the latter rather than the
former. Dang, what should I do? Is my code safe?

The code works now, so I guess all I had to do was change the
string to char and be done.