WTH!?
Why would this not work? Working with strings in Go is tricky. I
don't like it. There are too many footguns and in the learning
materials I have got so far there are not enough
conventions/strategies to handle these things. I don't like this.

Conclusion: accessing individual runes in a string is a tricky
thing and careful considerations need to be made when doing those
operations. What does Index even return? It returns a byte but it
does not tell me if it is the byte of a single-byte rune or a
multi-byte rune. I think this is a serious design flaw from the
Go designers. In every other language, using str[i] and len(str)
(or equivalent) is an operation that returns what you would expect.
This means that while in other languages a string is a sequence of
chars, in go a string is a sequence of bytes ... which is more
ambiguous and error prone. Just hide the complexity away! The UTF-8
format is not an excuse not to do so. If a dev wants an array of
bytes, then they can easily use bytes(str) and get it.

The program that I want to write in Go uses a lot of strings.
Should I stat with Go, or use something else?