Preparing for change of return type
I changed the tests to reflect the desired change in API. The
process was tedious but relatively straight forward ... except
that there are erros. More on that latter. I could have just
modified the function and see all tests pass, but I wanted to see
the tests fail first. It is a simple but critical step in TDD. It
is a way to make sure that tests pass because a satisfactory
implementation was made and not becuase the test is tautologically
passing.

Now, unto the erros. Take a look at the first test,
testIdenticalStrings. In this case the function must return
  (false, -1)

That is the reason why when I added thereIsAMatch into the
conditional that I negated it and made sure to AND it. ... but
that is incorrect. The conditional checks that the value is the
WRONG one, not the correct one. Therefore, instead of having
  !therIsAMismatch
I should have
  thereIsAMismatch == true

Also note that if either of the values is the wrong one, then the
test must fail. Therefore, instead of using && I should use ||.
The correct return value is
  thereIsAMismatch == false && mismatchLineNumber == -1
Since we seek to check if the return fails to satisfy these
conditions we must check if the result is equal to its negation.
That is, since
  return == acepted => PASS, then return != acepted => FAIL

The negation of the correct return value is
  thereIsAMismatch != false || mismatchLineNumber != -1

Notice how this test would have made it harder to get a test fail.
Meaning, there could be false positive. I thikng that is the worst
scenario. When tests keep failing despite the fact that the code
is right is frustrating, but at least you have good indicators that
something is wrong with the tests. In this case, the only way to
notice is to realize that the code is wrong but all tests are
passing. That is less likely for such a thing to happen. So yea, it
is really important to understand the problem domain and how to
write tests. This is another good incentive to use property-based
tests. Properties tend to be more rigurous and, due to their risk,
more caution is allocated when writting that kind of code.

Since I used this faulty test as the template for the modification
of all other tests, then I have to rewrite all the other tests.
This is something that I feared as I was writting the test. Since
there are so many tests, modifying them is time consuming and
risky. The whole point of writting tests is that you do not have
to be so careful when writting the implementation. The issue is
that you have to be extra careful when writting the tests. I think
this saves effort, since it costs less effort to be careful when
writting the tests than it is to be careful when writting code.
Proof of this is the fact that I already messed up big time twice
in the tests, and both times I recovered and finished in a timely
manner.

Still, this oversight cost me a lot of time. Should I work on
consolidating the tests into some sort of table or file? I can
give it a shot. But first I want to fix the tests. Then I can
quickly get the implementation right. Then I can change the tests.
That is a good strategy. Once you know that the program passes,
you can modify the tests. If the tests FAIL, you know that either
the test is wrong or the previous test was wrong.

Anyway, I better start these changes.