Change return type in impl
I changed the return types in the implementation. I thought I
would be confident about implementing the feature, but I got cold
feet. When should findLineMismatchOfTwoDifferentNonEmptyStrings
return false? That is when I realized what the name clearly
implies; which was my intention from the beggining - that is the
reason why I gave it that weird, long, and explicit name. The two
strings are not empty and they are different. Therefore, there
must be a line mismatch. The only case where the function should
return false is in the first if guard. All other cases it should
return true. Which means that I do not have to change the
implementation of the core of my algorithm ... which is a pleasant
surprise. I am keeping this just for the record, but I am going to
fix it on the next commit.

I had a really good gut feeling when I separated the
implementation into these two blocks. First I took care of simple
cases that were easy to understand and implement. That allowed me
to solve a more specific problem. This is a good example of how
property/case based programming is a goood approach to develop
algorithms that are easy to reason about and easy to maintain.