Fixed one test
The tests are passing but I know that is misleading. All the tests
need to be rewritten. Is TDD worth it? I mean, I think I got the
tests right, but did I really got the right? What if there is a
typo in the tests? What if there is another logical falacy? Is
there a way to tell? No. There isn't. The tests need to be
carefully crafted. As I see it, the idea is to be careful with the
tests instead of being careful with the implementation because the
former is less effort than the latter. But still, I feel uneasy.
Is all of this effort worth it? Is my code better due to all of
these measures?