Fixed tests
I am going to have to double check to make sure that the values
are correct. I will use gitg to check. Although, if the values
were wrong, since the impl is right, then there would be failing
tests, but there aren't. Therefore, it is likely that the values
are correct.

Right now it is clear thatt there should be more utility functions
for these things. Now all the functions are homogenous. I think
there is a lot of value to be gained from extracting all the test
cases into some format that is used in a for loop or something. I
still need to figure it out. I want to try and avoid using files
for now.