Midpoint
I am roughly halfway through the tests. This is tedious but it
doesn't look that bad. It is not trivial to read the test cases,
but it is easy to get used to. It does make things clearer.

I am going to have to change the descriptions of some tests. They
do not sound like a specification, but I have no idea how to 'fix'
that. I think it is ok, but I am open to other possibilities.

So far so good.