Better descriptions
I did my best, but I think that it is unsatisfactory. I don't know
how to properly name the tests. I am not sure if I am doing it as
a proper spec. Only the first two feel like a spec. The tests is a
shit show.

I also started the documentation of the package. It is worse than
what is going on with renaming tests.