Prep for Col Nums
Preparation to implement Colomn Number feature. I wanted to see
how difficult it would be to change the current code to implement
this feature. I struggled to figure out what the issue was with
line 57 in case five, so I decided to go ahead and impl this
feature.

It was tedious but straightforward to change the tests. I used an
Elixir interactive terminal to confirm all of the values.

It was also simple to create a failing impl that matched the new
signature of the function.

While I was modifying the test entry point I also added a report
with information regarding how many test were run, how many failed,
and the percentage. Just for diagnostics.