New fn with tests
I decided to implement the column mismatch feature as its own
function. I think it could be useful for finding stowaway chars
but I am not sure yet.

This is the stuff that has always pissed me of about testing. This idea that the only things that
need to be tested are public functions. I am planning on implementing findColumnMismatch. I know
that I am going to have to work with individual runes(aka chars). Therefore, there is a chance that
I am going to get things wrong because I've never worked on something like this before. Thus, I
want to write a couple of tests just to be sure that I am doing it right. There is no reason for
this function to be public ... but I cannot test it if it is not public. I find this limitation
infuriating. The only option I have is to make it public. It is a useful fn, so I see no reason why
not to, but still, these constrains are forcing me to change the way that I aproach testing my
software.

I was not sure if I could move the struct definition within a function. It worked. I am happy. It
makes the code a little hard to read, but it makes sense. The benefit of keeping things local makes
things a lot safer. I think it might also help with the memory consumption of the program, right?
Nah. Constants are constants. It goes into the .data section.

Regardless. I made the function public. I managed to get the test going and a failing
implementation.

As I was working on TestFindColumnMismatch I realized I made a mistake. The column cound is from the
begginging of the new line, not the begginging of the string! This is the third time (right?) that
something like this happens. It is the third strike. This is definetely not sustainable. On the
other two ocasions I was a little confused/mentaly tired and I noticed the mistakes right away. This
one is different. I was sharp when I did those test. Not only did I counted the columns by hand, I
doubled checked it in an IEx terminal. This is a terrible oversight. I would have noticed during
debugging, when my tests failed despite getting it right. This is possible in this case because the
problem is simple. But what if the problem had been more complicated? Here definetely a
property-based approach is safer.

The generative tests is where you put most of your effort. You make sure you get it right. You
accept the chance that the property test is wrong. You make some value-based examples. You put some
effort into making sure that they are right. You accept the chance that the tests are wrong due to
human error. The property-based test is the heavy lifter in the verification process. It makes sure
that the implementation is correct. In the event that the implementation and the property are
correct but the value bassed isn't, then the impl + debugger can be used to fix the value based
examples. The job of the value based examples is to detect an issue with the property-based test.
These are three players. So long as two are correct, you can use the two to fix the third.

Reaching that stability is the tricky and critical part of software verification. It is about
reaching a level of confidence about the state of your software. Reaching it is the tricky and
expensive part. You really have to be confident that your system is doing what is intended. That you
have done your due dilligence and that you have peace of mind. Once you have reached that point, you
modify these three elements one at a time. That is, if you want to add/modify value-based tests
(without adding new features of course), you freeze the other two. If your tests fail after the
modifications, then you did something wrong with your new tests.
freeze the other two. You make the changes

One critical component of this process is gradual increments. If your app needs to do 10 things, you
set up these three pilars for only one of those things. And you take your sweet-ass time making sure
everything works. Then and only then do you attempt to implement the next feature.

Ok. So what to do about the current situation? I am confident that the tests for FindColumnMismatch
are correct. Why? Because the problem is simple. I counted them and recounted them. They are
correct. Ok. We need a second pilar. I could implement the code and then go with the property based
but that clearly is not the way. I need the property first. No implementation can be trusted without
tests. Thus, this is not a situation of the chicken and the egg. The tests always precede an
implementation. In this case I could do a single implementation, since I trust the test so much. It
works because the feature is really simple and the tests are clearly correct. Had the problem been
more complex then I would go with the generative approach without question. But I am salty about the
mistake that I made. Therefore, I am going to go with the property test next. After that I am going
to attempt to implement the feature.

... I am having trouble developing the feature. I think I ran into the issue of the only way to test
the implementation is by implementing it twice. Let's think. FindColumnMismatch returns an int. That
int signals the point at which two strings differ. That means that before that point the two strings
are equal and on and beyond that point the strings are different. That is:

  i := FindColumnMismatch(left, right)
  frontLeft := left[0 : i]
  frontRight := right[0 : i]

  frontLeft == frontRight

This property is great! But, by chance, it happens that right now in VS Code the test right in front
of me is the 'substring' case. I am not sure how this property would fit in those circumstances. But
that has to do with array slicing and not with the issue that I have with strings.

I was thinking of reading about what happens when you access and array out of bound, but I was
filled with dread. I knew that if I went down that route I was going to find a lot of information
that was not related to the subject, that was not going to help me, that was going to confuse me,
and that I was going to have to give it a try anyway. So I just went ahead and did a simple
experiment:

func main() {
	left := "same sentence"
	right := "same sentenc3 mismatch"
	i := 14

	newLeft := left[0:i]
	newRight := right[0:i]

	fmt.Println(newLeft)
	fmt.Println(newRight)
}

which resulted in:

panic: runtime error: slice bounds out of range [:14] with length 13

It is a little disappointing, since both left and right are essentially constants. Why did it not
detect it at compile time? Well. Anyway. No safe access to slices.

Does this split by bytes or by runes?

func main() {
	left := "same sentence"
	right := "same sentence mismatch"
	i := 14

  var newLeft string
  var newRight string

	if i < len(left) {
		newLeft = left[0:i]
	} else {
		newLeft = left
	}

	if i < len(right) {
		newRight = right[0:i]
	} else {
		newRight = right
	}

	fmt.Println(newLeft)
	fmt.Println(newRight)
}

Yes. It works as expected. Maybe I am overthinking it when it comes to runes. Maybe len() is the
only unsafe operation that can be done with strings. Regardless. This is a good prototype for the
property based test. Yep, this is good enough to proceed.

-------------
Before I do anything else, I must fix the tests for FindLineMismatch. Otherwise I might forget and
it will be a dissaster.