fixed flagged error
I forgot to mention, I noticed this mistake when I was working on
"A single newline char difference at the end is a mismatch." I was
trying to figure out how to handle an empty line at the end when I
remebered not to considering that case before. It was dumb luck.
This is the reason why values without properties are not robust.
But as in the case made a couple of commits ago, sometimes
properties are hard to design, not all properties are equally
useful, and depending on the complexity, there is a chance for the
properties to have bugs themselves.

Is testing worth it?