Fixed tests
What is embarasing about this is that the last test is correct.
How is it that you notice it for the weirdest case but not for the
previous two? This case is in sharp contrast to the case in the
last commit. IT WAS DUMB LUCK! The mind can be finicky like that.
One day you are full throttle, other days the gears slip. Days,
more like hours. It is constantly fluctuating. More points in
favor to property-based testing.

Is testing worth it?