Failed property
My idea was to use two loops to create two random strings. Here is
the thing though: what are the odds that the two strings would
mismatch on the first rune? For N symbols in the rune set
(ignoring repeats - repeating the whitespace was a mistake. I will
talk about it latter), the prob of the two runes being different
on the first char is (N - 1)/N. Meaning, really high. This can be
extended to the second and third rune. Meaning, most of the tests
are going to return less than 4. That means, I am doing a lot of
work without really creating meaningful examples.

What I really want out of my generator is to produce tests like
the ones that I am writting, but giving the generator more leeway
and resources. This is a good point in favor of writing the value
tests first. This gives the dev a good idea of what the output of
the generator should look like.

Ok. So what now? I want my generator to output something like what
I am writing. Instead of doing it one char at a time, it is best
to do it one word at a time. It is hard to describe. I have it
cooking in my head. The only way tog et it out is with code.