Record changes
I am changing things around and I want to keep a record.

As I was writting the for loop in L182 I realized that there was
no need to keep the logic of how big a word should be outside of
the createWord fn. I wish I had kept a record of that too. I moved
the logic in and now I have to deal with the fact that I am
working with an array isntead of a slice. I just want to keep
track of the changes.

I should have done so sooner.