Extraction
I extracted the logic to generate string examples. I also
fixed a critical typo whe using strings.Join.

The git algorithms is doing something weird when it comes to
diffing, so the changes won't show up properly. The only way to
see it is to check the two file versions out and compare.