First trial of property test
I got a panic:

  panic: runtime error: slice bounds out of range [:-10] [recovered]

I have to sanitice the index returned by the function. This test
cannot handle that case. This reminds me that I wanted to use some
probability to test equal and unequal strings. I need to be able
to handle all cases. I just noticed that my value tests are missing
when one of the strings is empty. I need to add that.

For now, I will focus just on saniticing the index given by the
function.

I am planning on having two different functions. One that tests
equal strings and one that tests unequal strings. That means that
I have to guarantee that the strings being generated are different.
I have to add a conditional to test if that is true.