Added more value tests
I discussed in a previous commit that there were some cases that I
forgot to create in the value based tests.

I also realized that there is no need to check the case where both
strings are equal, since that case is never going to be reached.
But I remembered considering that case. Maybe in the commits there
is evidence of me creating and deleting that case. In the event
that I didn't, this would indicate that I knew this implicitly.
That is, I was aware that both strings were not equal. But I did
not write it anywere! Meaning, there are 'properties' of the code
that live only in my head and are not 'explicitely' or 'actively'
known. That is problematic for all kind of reasons, but I think it
is something integral about software development. No matter how
hard we try, there are things about the code that are not in the
code. Or at least not explicitly. This is something that
compromises the understandability of the code and it is something
that must be actively fought against.

I am making a big deal out of this not because it is a big deal but
because it is the first time that I was able to document it. It is
a good example.

While I was typing this, I realized that the sanitation for the
index was not needed. That is right. I almost forgot. In the
previous commit I did mention that I had to account for when both
strings were equal, returning -1. And that I had to sanitize the
returned value. The returned value was negative because I
hardcoded a value of -10. I did so with the anticipation of
returning -1 once it was implemented. So which of teh two was it?
Had I been aware of that posibility, I would have added the test
case. I think it was both. I did the failing impl first. I just
copied what I did with the previous failing impl. But as I was
writing the tests I must have known that the two strings being
equal was not going to happen.

Regardless, this made me realize that I did not have to do the
sanitation. But then I realized that the function must not return
zero either. The sanitation does take that into consideration but
I got concerned for some reason. That there is something in the
code that I wrote that could return zero. It was rand.Intn, right?
I am confused now. I was concerned about something, but I can't
remember it.