aaaahhh chingaohhh
Wha happened!?!?

What happened to the rest of the tests? I guess I don't understand
FailNow after all. It shuts down the whole thing. But that didn't
happen to the other tests?

I have this feeling that it is like in Elixir, where if a child
implotes the parent could implote in tandem. Like, the t that I
am passing to the property is the parent and I am shutting it
down, shutting the whole thing down.

I have to options. 1) Refactor to get it working with onlt Fail.
That one is going to be easy and it does not compromise the
understandability of the code. 2) use t.Run to create a new
child process that I can shut off without shutting off the whole
thing. I want to give the second a try because I want to learn.
If it fails, I can always back up to 1).