Loop the property test
I wrote a loop for the property-based tests. I was looking at the
output and it looks correct. What I am concerned about is whether
it works for the case when the word cound to be selected to be
a mismatch is 1. That is, can the generator produce examples that
differ by the first word? I had this fear before but I could not
express it in words. I askedmy self, should I add 1 to the range?
Only now does it surface.

Looking back at the code, it is correct. The difference will never
occur in the first word. I guess I wanted to guarantee that there
would be at least one mismatch? Maybe it was that. I don't know.
But mismatchAtThisWordCount is never zero. Wait.

If mismatchAtThisWordCount == 1, then mismatchAtThisWordCount - 1 == 0.
That means that the first for block is not executed. Then we do
the logic to make a mismatched word. That var is counting using
natural numbers.

I have no clue why I did that. And I did it today! I guess it was
too early into the development of the function, I was describing
the problem as it was in my head and not as a gear in an algorithm.

I guess I was also hoping to use the variable to indicate where
the difference is supposed to be to make it even easier to find
the mismatch. I think there is value in that. I have a faint
memory of dropping that idea because I could not use that variable
to pinpoint the exact column number. Still, having the word count
is useful. It is actionable. I guess I can still use it.

Nop, dumbass. Think again. It is not counting words. It is
indicating an index. When that var is one the mismatch happens on
the second word, not the first.

NOOO. Think again. It is couting the word number. Yes. The code
works. It does what is supposed to do.

Should I write a test to make sure that the generator is producing
what it is supposed to be producing? The word count should be the
sum of the two vars and the word count indicates where the
mismatch is. It is a simple property. I can test it. I should test
it. I want to test it. That is what really matters. That I want to
do it. A test that gives you doubt is not a test, it is a
liability. Yes, it is more work. But I need that peace of mind. I
need it. It is non-negotiable. It is critical. I need to make sure
my code is doing what it is supposed to be doing. And the cost is
going to be the cost. The faster you get at paying it, the better
off you are going be. The longer you hesitate, the more it is going
to cost you.