Confused about license
I am not sure how this license business works. I have tired to
learn, but all I find online is vague, contradicting, and always
followed by 'this is not legal advise' and the like. I understand
that law matters are delicate, but come on. Just how much do I
have to learn before I can release an open source software?

I was reading about all of this and I stumbled on a GitHub issue
where a Contributor License Agreement was mentioned. ... so there
is a different license agreement for contributions? It was also
mentioned an AUTHORS file and a CONTRIBUTORS file. The complexity
of these things is getting out of control. I need to find a
reliable/concrete learning material to learn about these things.
Otherwise I am going to be stuff in doubt and stress, and I am
going to fuck it up anyway.

Right now I am concerned. As I understand, the first comment in a
file is the description of a package. Then the stuff above the
package declaration is the rest of the docs for the package.
Something similar applies to the functions. So I am confused as to
what is going to happen if I put the license at the top. And I
can't find information about that specific subject. I can find
about doc strings in Go and I can find about licenses requiring to
be placed at the top of every file. It is really frustrating.

I am going to leave it as is because it is the best that I can do
for now. But I hate this situation. I feel like this all the time
when I am working with software. I can never be sure of anything.
Everything is ambiguous, has unlisted caveats, is a prototype, a
convention, is conflicted by X, Y, and Z. The whole thing is a
shitshow. Anhd it didn't have to be this way.

I am loosing my SHIT! The autoformater added a # in front of the
copyright. Why? Queries like "golang what does # mean" do not
yiled information on the subject. I don't know what to do.
I don't know how to configure the editor. Nothing.