I don't get it
I tried to see if it meant anything outside of a comment, but it
marks it as an error. I don't know what is this thing and it is
really bothering me.

There is no infor on the subject in https://tip.golang.org/doc/comment.
I have no clue what to do.