WTH
I was able to modify the generators.go file and not have a pound
but that was not working for generators_test.go. I noticed that
generators.go had a description of the package. So I decided to
add a comment in generators_test.go and the annoying thing
disapeared. Is that the reason why there is that automatic formating
thing? Why does it do this? What does # mean?