I am at my wits end
I found an example of how to place a license in a file in this
article: https://medium.com/@yanglyu5201/golang-development-documentation-ru-eb29cbb1083
It is the only example I have been able to find.

I imitated the example. At first the # sign was not inserted
automatically but then for some strange reason it is. I have no
idea what to do. I cannot find any information as to what that
means or why it is being done.