I figured it out
It is autoformating it to be a heading. I searched for the char '#'
in https://tip.golang.org/doc/comment and found that: "The # syntax
was added in Go 1.19. Before Go 1.19, headings were identified
implicitly by single-line paragraphs satisfying certain conditions,
most notably the lack of any terminating punctuation."

The Go formater thinks that it is a heading. What a nuisance. In
the same page, it does not mention that the doc for the package
needs to be the very first thing in the file. By the examples, it
looks like it is the first comment right above the package. Just
like with a function. With a function, I remember reading explicitly
that the first comment above the function is the doc comment. It
might be the case that it is like that here too. All I need to do
is separate the comments, contrary to the example I found online.