It worked
I guess I can move on. I hate computers. This is a clear example
of how having different types of comments would be very useful.
One for licenses and stuff like that. Other for configs and stuff,
like the !bin comment for shells and things like that, other for
dev comments(vanilla //) and one for oficial documentation.