I don't get it
It is fairly straight forward to make this a constant. Why can't I
make it a constant? I read online that it has to do with the fact
that in Go arrays cannot be declared immutable. I guess it comes
from this idea of speeding compilation times. Ok. It is a big deal
but ok.