Tests need to be rigurous
I was working on describing and creating the test for
CreateStringExamples. My strategy was to split the string into
words and compare the words to make sure there is only one word
mismatch. That is the only basic property that the generator has.
Then I noticed that both strings are always equal in length. I
started thinking if that was a good or a bad idea and concluded
that it was a bad idea.

I now realize that the issue was that I was thinking about writing
a test that tests my particular implementation. In a previous
commit I mentioned that I trusted myself to separate the details
of implementation and testing. That is, that I did not needed to
be so strict about writing the tests first before writing the
implementation. The fact that I am writing about this right now is
a testament to that discipline, but this shows how insinuous is
the influence of previous designs on current designs.

Right now what I need to do is to specify the kind of output that
the generator needs to output and write test examples for it. So
far I have been writing a test for a test for a test. When does it
end? It ends when I am confident that whatever I am coding is doing
what it is supposed to be doing. Yes, it is subjective and just as
flawed as any other programming (human) endeavour. All I need to do
is keep testing until I am confident that the code does what is
supposed to do. I think this package is going to be a lot easier
to test. I am getting close. The more I invest, the more solid my
implementation, and the faster and easier it will be to modidy in
the future.