Start of new test specification
This is going to be a lot harder than I expected. This generator
is stochastic. I won't be able to guarantee that any one run will
produce any one of these results. Then, how do I write the tests?
I was thinking of using a similar strategy I used in
FindLineMismatch where I create multiple specific implementations
and I choose one as appropriate. In this case I could implement
multiple functions that implement one specific feature. Then I can
test each one individually. Something like this:

func HasBegginingWhitespace(num int, words []string) []string
func HasTrailingWhitespace(num int, words []string) []string
func EqualWords(num int, words []string) []string
func DiffWords(words []string) []string

The first function generates a string with the right number of
whitespaces and adds it to the list of strings to be joined.

Similarly for the second function.

The third function generates a list of words to be used by both
strings.

The fourth function add single words that differ at at least
one character.

This kind of reminds me to the builder pattern. I am thinking of
using a struct that takes all of this information and passes it
to a builder that creates the desired string.

Then I can do:

var output []string
HasBegginingWhitespace(2,
    EqualWords(4,
        DiffWords(output)))

Which looks really ugly, but I can verify that these functions
implement the correct thing. Again, it is all about confidence.

But that implies creating more and more public functions ... and
I don't think that is the right way to go. That might be a good
idea if I want to work on a bunch of atomic components to be used
in multiple generators. This might be a good idea for a generator
library. But that is not what I am programming right now.

I think there is a better approach. For this kind of stochastic
generator there is no way to know what particular instance is
going to output. But what we can do is run it many times and check
if a specific instance happened at least once. If it happened
once, that means that the generator is capable of producing it
(duh!). What I am trying to say is that all we need is to check if
a given characteristic happened once.

In that case, instead of using a fixed run number, we can set it
as an execution limit, and instead halt the test once all cases
have been generated. A regex is a simple (yes, regexes are simple)
and efficient way of checking for properties of strings. It is
something that fits the current type of tests that I am writting.
I think it would be a lot easier to specify the types of outputs
that this generator should procude using regexes than anything else.
It would be faster than the builder pattern. The builder pattern
might be a good idea too. Are there any libraries that do this in
Go? Probably.

I did a quick search, there is suss, rapid, check (Scala?), to name
a few. It isn't a big space, I think. When it comes to testing, I
think it is possible to do it in other langs too. Maybe property
based tests in Go are made using another lang/framework pair.

Go has regexp in its stdlib.

I need to sleep over it.